requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# pysbpl/setup.py: 2,7
Cython == 0.29.21

# pysbpl/pysbpl/map_util.py: 4
PyYAML == 5.3.1

# pysbpl/pysbpl/gen_mprim.py: 11
# pysbpl/pysbpl/plot.py: 3,4
matplotlib == 3.2.2

# pysbpl/pysbpl/gen_mprim.py: 11
# pysbpl/pysbpl/planner.py: 2
# pysbpl/pysbpl/plot.py: 2
numpy == 1.19.2

# pysbpl/pysbpl/map_util.py: 3
opencv_contrib_python == 3.4.0.12

# pysbpl/setup.py: 1
setuptools == 20.7.0
